fix: useGetIndustryOptions hook added, fixed validations
